+++ test/unit.sh
@@ -101,6 +101,7 @@ runUnitTest $D/issues/json_env_passing_4080.mocha.js
 
 runUnitTest $D/issue_6106_windows_home.mocha.js
 runUnitTest $D/issue_5990_bun_substring_match.mocha.js
+runUnitTest $D/interpreter_runtime.mocha.js
 runUnitTest $D/issue_6089_prototype_pollution.mocha.js
 runUnitTest $D/issue_6075_redos.mocha.js
 runUnitTest $D/issue_6073_object_env.mocha.js